A working version of std_meshadapt_solver
We still need to close the adaptive loop and the main barrier there
seems to be the conversion of ParMesh to ParPumiMesh. The operation
works fine in the corresponding C++ code by using dynamic cast. But a
wrapper call that only does a dynamic cast seems to not be able to do
the job!